# Optimize IntMap.alter using unboxed sums.