@m-renaud 2018-02-01
I'll add more fine grained benchmarks if I get the chance tomorrow. Sending out for feedback on the code.