@m-renaud 2018-02-01
So, as is always the case, using unboxed sums makes some operations faster, and others *a lot* slower :(

**Before:**
```
benchmarking alter absent
time                 283.1 μs   (262.8 μs .. 298.1 μs)
                     0.982 R²   (0.977 R² .. 0.991 R²)
mean                 270.9 μs   (262.3 μs .. 283.0 μs)
std dev              32.15 μs   (24.01 μs .. 45.31 μs)
variance introduced by outliers: 84% (severely inflated)
               
benchmarking alter insert
time                 277.1 μs   (273.3 μs .. 282.3 μs)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 275.7 μs   (271.7 μs .. 281.7 μs)
std dev              16.49 μs   (11.93 μs .. 27.52 μs)
variance introduced by outliers: 56% (severely inflated)
               
benchmarking alter update
time                 280.3 μs   (276.7 μs .. 284.3 μs)
                     0.998 R²   (0.998 R² .. 0.999 R²)
mean                 280.5 μs   (277.7 μs .. 284.5 μs)
std dev              11.21 μs   (8.620 μs .. 15.42 μs)
variance introduced by outliers: 36% (moderately inflated)
               
benchmarking alter delete
time                 282.5 μs   (274.8 μs .. 289.7 μs)
                     0.996 R²   (0.995 R² .. 0.998 R²)
mean                 281.1 μs   (276.0 μs .. 290.9 μs)
std dev              21.75 μs   (12.40 μs .. 34.62 μs)
variance introduced by outliers: 68% (severely inflated)
               
benchmarking alter delete absent
time                 243.8 μs   (240.8 μs .. 246.9 μs)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 241.7 μs   (239.1 μs .. 245.8 μs)
std dev              11.08 μs   (8.402 μs .. 14.84 μs)
variance introduced by outliers: 43% (moderately inflated)
```

**After:**
```
benchmarking alter absent (-16%)
time                 237.9 μs   (234.0 μs .. 241.8 μs)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 237.0 μs   (232.9 μs .. 246.5 μs)
std dev              20.16 μs   (7.155 μs .. 32.95 μs)
variance introduced by outliers: 74% (severely inflated)
               
benchmarking alter insert (+290%)
time                 805.2 μs   (783.5 μs .. 828.1 μs)
                     0.990 R²   (0.979 R² .. 0.996 R²)
mean                 822.7 μs   (799.7 μs .. 872.8 μs)
std dev              110.8 μs   (59.63 μs .. 192.9 μs)
variance introduced by outliers: 84% (severely inflated)
               
benchmarking alter update (+250%)
time                 705.5 μs   (699.2 μs .. 712.7 μs)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 703.5 μs   (695.1 μs .. 716.0 μs)
std dev              34.66 μs   (25.56 μs .. 49.30 μs)
variance introduced by outliers: 41% (moderately inflated)
               
benchmarking alter delete (+255%)
time                 720.1 μs   (684.8 μs .. 766.0 μs)
                     0.987 R²   (0.978 R² .. 0.998 R²)
mean                 684.8 μs   (674.2 μs .. 702.9 μs)
std dev              46.01 μs   (28.89 μs .. 72.45 μs)
variance introduced by outliers: 57% (severely inflated)
               
benchmarking alter delete absent (-8%)
time                 225.4 μs   (223.0 μs .. 227.6 μs)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 225.1 μs   (222.4 μs .. 231.9 μs)
std dev              13.72 μs   (6.353 μs .. 26.03 μs)
variance introduced by outliers: 58% (severely inflated)
```

It looks like in cases where we need to modify the spine anyways the extra pattern matches (despite being unboxed sums) on every `Bin` have quite a bit of overhead.