@treeowl 2018-02-05
Take a look at https://github.com/treeowl/containers/commit/6f6fa02c84a4839413015bf7d52f8dc18bc33157. That's not very well organized yet, and doesn't deal with the lazy version, but it shows how the strict version can be made more eager, how pattern synonyms can cut down dramatically on the noise, and how we can "unbox" the `Maybe`s consumed and produced by the passed function. Feel free to incorporate those ideas into your branch if you wish.