@treeowl 2018-02-05
We need to benchmark the unboxed sums version against one that just uses pointer equality on each recursive call.