@sjakobi 2020-08-01
What's the status here, @m-renaud?

In light of https://github.com/haskell/containers/issues/538 it would be nice to make progress on this.