@sjakobi 2021-09-15
> Shall I try to take this up and finish it

That would be great! :)

> anyone know what there is to do left?

I think the minimum is to fix any strictness issues such as https://github.com/haskell/containers/pull/523#discussion_r165466312.

If the benchmarks show good improvements at this stage, I think this should be good to merge.

Adopting some of noise reduction measures suggested in https://github.com/treeowl/containers/commit/6f6fa02c84a4839413015bf7d52f8dc18bc33157 probably wouldn't hurt though.

Regarding further optimization ideas, I think those could be left for further work.