@Kleidukos 2022-11-02
What is the status on this one? Are those number still the same on GHC 9.2?