@treeowl 2022-11-02
> What is the status on this one? Are these benchmarks numbers still the same on GHC 9.2?

Has something changed there?