@Kleidukos 2022-11-02
No unfortunately I'm quite committed on Haddock and other stuff for the time being. How's the maintainer structure for `containers` going btw? Need any help on this front?